be/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
pydantic==2.5.0
fpdf2==2.7.4
python-multipart==0.0.6